requirements.txt
# Force Pydantic v2 for LangChain
pydantic>=2.7.4

# Force a version of FastAPI that works with Pydantic v2
fastapi>=0.100.0

# Web Server
uvicorn[standard]

# AWS & LangChain
boto3
langchain
langchain-core
langchain-aws
langchain-community

# Tools & Data
pypdf
faiss-cpu
google-search-results
python-dotenv
langgraph